+++ helper_function.h
@@ -8,15 +8,15 @@
 #include <string.h>
 #include <math.h>
 #include <time.h>    
-
+#include "std_types.h"
 #define NUM_ITERATIONS 17
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
 #endif
 
-double generate_normal_random(double mean, double stddev);
-double rand_double_range(double min, double max);
+float generate_normal_random(float mean, float stddev);
+uint32_t rand_double_range(uint32_t min, uint32_t max);
 
 
 #endif /* HELPER_FUNCTION_H_ */
